+++ Braintree.Tests/CreditCardTest.cs
@@ -523,6 +523,29 @@ namespace Braintree.Tests
         }
 
         [Test]
+        public void verifyValidCreditCardSpecifyingMerhantAccount()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = "5105105105105100",
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true,
+                    VerificationMerchantAccountId = MerchantAccount.NON_DEFAULT_MERCHANT_ACCOUNT_ID
+                }
+            };
+
+            Result<CreditCard> result = gateway.CreditCard.Create(request);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(MerchantAccount.NON_DEFAULT_MERCHANT_ACCOUNT_ID, result.CreditCardVerification.MerchantAccountId);
+        }
+
+        [Test]
         public void verifyInvalidCreditCard()
         {
             Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;